[#218] 서치바 UI 애니메이션 및 프레젠테이션 로직 수정, 스크롤 시 키보드 숨기기
